+++ src/hadi1992.jl
@@ -85,15 +85,23 @@ function hadi1992(multivariateData::Array{Float64,2}; alpha = 0.05)
     # Step 0
     meds = coordinatwisemedians(multivariateData)
     Sm = (1.0 / (n - 1.0)) * (multivariateData .- meds')' * (multivariateData .- meds')
-    mah0 =
-        diag(mahalanobisSquaredMatrix(multivariateData, meanvector = meds, covmatrix = Sm))
+    
+    msm1 = mahalanobisSquaredMatrix(multivariateData, meanvector = meds, covmatrix = Sm)
+    @assert !isnothing(msm1)
+
+    mah0 = diag(msm1)
+    
     ordering_indices_mah0 = sortperm(mah0)
     best_indices_mah0 = ordering_indices_mah0[1:h]
     starting_data = multivariateData[best_indices_mah0, :]
 
     Cv = coordinatwisemedians(starting_data)
     Sv = (1.0 / (h - 1.0)) * (starting_data .- Cv')' * (starting_data .- Cv')
-    mah1 = diag(mahalanobisSquaredMatrix(multivariateData, meanvector = Cv, covmatrix = Sv))
+    
+    msm2 = mahalanobisSquaredMatrix(multivariateData, meanvector = Cv, covmatrix = Sv)
+    @assert !isnothing(msm2)
+
+    mah1 = diag(msm2)
     ordering_indices_mah1 = sortperm(mah1)
 
     r = p + 1
@@ -113,23 +121,18 @@ function hadi1992(multivariateData::Array{Float64,2}; alpha = 0.05)
         if det(cfactor * Sb) == 0
             @info "singular Sb case"
             newSb = hadi1992_handle_singularity(cfactor * Sb)
-            mah1 = diag(
-                mahalanobisSquaredMatrix(
-                    multivariateData,
-                    meanvector = Cb,
-                    covmatrix = newSb,
-                ),
-            )
+
+            msm3 = mahalanobisSquaredMatrix(multivariateData, meanvector = Cb, covmatrix = newSb,)
+            @assert !isnothing(msm3)
+            
+            mah1 = diag(msm3)
+            
             ordering_indices_mah1 = sortperm(mah1)
             basic_subset_indices = ordering_indices_mah1[1:r]
         else
-            mah1 = diag(
-                mahalanobisSquaredMatrix(
-                    multivariateData,
-                    meanvector = Cb,
-                    covmatrix = (cfactor * Sb),
-                ),
-            )
+            msm4 = mahalanobisSquaredMatrix(multivariateData, meanvector = Cb, covmatrix = (cfactor * Sb))
+            @assert !isnothing(msm4)
+            mah1 = diag(msm4)
             ordering_indices_mah1 = sortperm(mah1)
             basic_subset_indices = ordering_indices_mah1[1:r]
         end