+++ src/hadi1994.jl
@@ -58,15 +58,21 @@ function hadi1994(multivariateData::Array{Float64,2}; alpha = 0.05)
 
     meds = coordinatwisemedians(multivariateData)
     Sm = (1.0 / (n - 1.0)) * (multivariateData .- meds')' * (multivariateData .- meds')
-    mah0 =
-        diag(mahalanobisSquaredMatrix(multivariateData, meanvector = meds, covmatrix = Sm))
+    
+    msm1 = mahalanobisSquaredMatrix(multivariateData, meanvector = meds, covmatrix = Sm)
+    @assert !isnothing(msm1)
+    
+    mah0 = diag(msm1)
+
     ordering_indices_mah0 = sortperm(mah0)
     best_indices_mah0 = ordering_indices_mah0[1:h]
     starting_data = multivariateData[best_indices_mah0, :]
 
     Cv = coordinatwisemedians(starting_data)
     Sv = (1.0 / (h - 1.0)) * (starting_data .- Cv')' * (starting_data .- Cv')
-    mah1 = diag(mahalanobisSquaredMatrix(multivariateData, meanvector = Cv, covmatrix = Sv))
+    msm2 = mahalanobisSquaredMatrix(multivariateData, meanvector = Cv, covmatrix = Sv)
+    @assert !isnothing(msm2)
+    mah1 = diag(msm2)
     ordering_indices_mah1 = sortperm(mah1)
 
     r = p + 1
@@ -97,13 +103,14 @@ function hadi1994(multivariateData::Array{Float64,2}; alpha = 0.05)
             end
         end
 
-        mah1 = diag(
-            mahalanobisSquaredMatrix(
-                multivariateData,
-                meanvector = Cb,
-                covmatrix = (cfactor * Sb),
-            ),
+        msm3 = mahalanobisSquaredMatrix(
+            multivariateData,
+            meanvector = Cb,
+            covmatrix = (cfactor * Sb),
         )
+        @assert !isnothing(msm3)
+
+        mah1 = diag(msm3)
 
         ordering_indices_mah1 = sortperm(mah1)
         basic_subset_indices = ordering_indices_mah1[1:r]